Merged John's enhancements to TimeTrace
From the main RAMCloud repository (commit 6485430):
"""
You can now provide up to 4 integer arguments in each call to
TimeTrace::record, and the event string is now a printf-style
format string that can reference the arguments (the formatting
is only invoked if/when the trace is printed).

In addition, added prefetching to reduce the likelihood of cache
misses. The net of this is that the cost of TimeTrace::record
increase from 13.4ns to about 15.9ns.
"""